requirements.txt
# requirements.txt

asgiref==3.9.1
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
Markdown==3.8.2
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
